Move OpenStack backend files to help support multiple cloud platforms
- OpenStack files have been moved to controllers/api/openstack
- server.py has been modified - imports modules based on which cloud platform is being used
- Set which cloud is used by changing cloudPlatform in config/global.conf
- Slight changes to importing from getFunctions.py - due to change of location of files
- OpenNebula currently logs in but doesn't have functionality
- OpenStack works the same as before this commit
