+++ controllers/api/openstack/flavors.py
