+++ controllers/api/openstack/templatelist.py
