+++ server.py
@@ -11,13 +11,6 @@ import subprocess
 from subprocess import Popen
 
 from controllers.home import Home
-from controllers.machines import Machines
-from controllers.api.vm import VM
-from controllers.api.quota import Quota
-from controllers.api.user import User
-from controllers.api.templatelist import TemplateList
-from controllers.api.flavors import Flavors
-from controllers.api.projects import Projects
 
 cherrypy.config.update("config/default.conf")
 try:
@@ -38,21 +31,42 @@ badwords = f.readlines()
 cherrypy.config.badwords = [w.strip() for w in badwords]
 f.close()
 
+cloudPlatform = cherrypy.config.get("cloudPlatform")
+
 # PAGES
 home = Home()
-home.machines = Machines()
+if cloudPlatform == "opennebula":
+    from controllers.api.opennebula.machines import Machines
+if cloudPlatform == "openstack":
+    from controllers.api.openstack.machines import Machines
 
+home.machines = Machines()
 cherrypy.tree.mount(home, "/", "config/pages.conf")
 
 # API
 api = lambda:None
+
+if cloudPlatform == "opennebula":
+    from controllers.api.opennebula.vm import VM
+    from controllers.api.opennebula.quota import Quota
+    from controllers.api.opennebula.user import User
+    from controllers.api.opennebula.templatelist import TemplateList
+
+if cloudPlatform == "openstack":
+    from controllers.api.openstack.vm import VM
+    from controllers.api.openstack.quota import Quota
+    from controllers.api.openstack.user import User
+    from controllers.api.openstack.templatelist import TemplateList
+    from controllers.api.openstack.flavors import Flavors
+    from controllers.api.openstack.projects import Projects
+
+    api.flavors = Flavors()
+    api.projects = Projects()
+
 api.vm = VM()
 api.quota = Quota()
 api.user = User()
 api.templatelist = TemplateList()
-api.flavors = Flavors()
-api.projects = Projects()
-
 cherrypy.tree.mount(api, "/api", "config/api.conf")
 
 # Launch websockify for NoVNC
